refactor(packages): refactor packages structure
- rename all the export files and page(screen) files
- create a showroom package for all the widgets' example
- refactor the structure of packages(Please check the google drive file for more detail)
