+++ packages/ui_library/fonts/icon/UIcons.ttf
