+++ packages/ui_library/lib/icons/icons_export.dart
@@ -1,0 +1,2 @@
+export 'u_icons/u_icon.dart';
+export 'u_icons/models/models_export.dart';
