+++ packages/ui_library/lib/icons/u_icons/models/models_export.dart
@@ -1,0 +1,2 @@
+export 'u_icon_data.dart';
+export 'u_icons.dart';
