+++ packages/ui_library/lib/icons/u_icons/models/u_icon_data.dart
@@ -1,4 +1,4 @@
-part of 'u_icon.dart';
+import 'package:flutter/material.dart';
 
 /// Class used to create standard UIcons.
 ///
