+++ packages/ui_library/lib/ui_library_export.dart
@@ -1,0 +1,4 @@
+library ui_library;
+
+export 'const/const_export.dart';
+export 'icons/icons_export.dart';
