+++ packages/ui_library/pubspec.yaml
@@ -2,6 +2,7 @@ name: ui_library
 description: A new Flutter package project.
 version: 0.0.1
 homepage:
+publish_to: none
 
 environment:
   sdk: ">=2.16.2 <3.0.0"
