+++ packages/ui_showroom/lib/ui_showroom_export.dart
@@ -1,0 +1,3 @@
+library ui_showroom;
+
+export 'ui_pages/ui_pages_export.dart';
