+++ pubspec.yaml
@@ -15,7 +15,10 @@ dependencies:
     sdk: flutter
   intl: ^0.17.0
   ui_library:
-    path: 'ui_library'
+    path: 'packages/ui_library'
+  # delete showroom when exporting production app
+  ui_showroom:
+    path: 'packages/ui_showroom'
 
 dev_dependencies:
   bloc_test: ^9.0.3
@@ -30,4 +33,4 @@ flutter:
   fonts:
     - family: UIcons
       fonts:
-        - asset: ui_library/fonts/icon/UIcons.ttf
+        - asset: packages/ui_library/fonts/icon/UIcons.ttf
