--- ui_library/lib/core/all.dart
@@ -1,1 +1,0 @@
-export './icons/icons.dart';
