--- ui_library/lib/examples/all.dart
@@ -1,1 +1,0 @@
-export 'icons/icons.dart';
