--- ui_library/lib/ui_library.dart
@@ -1,5 +1,0 @@
-library ui_library;
-
-export 'widgets/all.dart';
-export 'core/all.dart';
-export 'examples/showroom.dart';
