--- ui_library/lib/widgets/all.dart
@@ -1,1 +1,0 @@
-export 'icon/icon.dart';
