--- ui_library/lib/widgets/icon/icon.dart
@@ -1,1 +1,0 @@
-export 'u_icon.dart';
